--- a/setup.cfg
+++ b/setup.cfg
@@ -4,7 +4,8 @@ exclude =
 	venv*
 	docs
 	build
-ignore = E704,W503,PYD002
+ignore = E704,W503,PYD002,TC003,TC006
 per-file-ignores =
     # The traces have to be formatted this way for the tests.
     tests/expected_traces.py: E501
+type-checking-pydantic-enabled = True