### What I did

before:

```
In [1]: %time import ape_foundry
CPU times: user 1.87 s, sys: 219 ms, total: 2.09 s
Wall time: 2.23 s
```

after:

```
%time import ape_foundry
CPU times: user 8.19 ms, sys: 3.58 ms, total: 11.8 ms
Wall time: 12 ms
```

### How I did it

### How to verify it

### Checklist

- [ ] Passes all linting checks (pre-commit and CI jobs)
- [ ] New test cases have been added and are passing
- [ ] Documentation has been updated
- [ ] PR title follows [Conventional Commit](https://www.conventionalcommits.org/en/v1.0.0/) standard (will be automatically included in the changelog)